+++ spec/spec_helper.rb
@@ -9,6 +9,7 @@ $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
   rake
   logger
   timecop
+  byebug
 ).each do |lib|
   require lib
 end